add cell-guide e2e test file
