+++ client/__tests__/e2e/e2e.test.ts
@@ -47,9 +47,7 @@ import {
   checkGenesetDescription,
   assertUndoRedo,
   snapshotTestGraph,
-  expandGeneset,
   getAllCategories,
-  expandMarkerGeneSetsHeader,
 } from "./cellxgeneActions";
 
 import { datasets } from "./data";
@@ -60,7 +58,6 @@ import {
   pageURLTruncate,
   testURL,
   pageURLSpatial,
-  pageURLCellGuide,
 } from "../common/constants";
 import { goToPage } from "../util/helpers";
 
@@ -104,118 +101,6 @@ const testURLs = {
   [DATASET]: testURL,
   "super-cool-spatial.cxg": pageURLSpatial,
 };
-
-describe(`Testing CellGuideCXG at ${pageURLCellGuide}`, () => {
-  test("page launched", async ({ page }, testInfo) => {
-    await goToPage(page, pageURLCellGuide);
-
-    const element = await page.getByTestId("header").innerHTML();
-
-    expect(element).toMatchSnapshot();
-
-    await snapshotTestGraph(page, testInfo);
-  });
-  test("assert absence of 'Standard Categories' and 'Author Categories'", async ({
-    page,
-  }) => {
-    await goToPage(page, pageURLCellGuide);
-
-    // Check for the absence of 'Standard Categories'
-    const standardCategories = await page
-      .locator("text='Standard Categories'")
-      .count();
-    expect(standardCategories).toBe(0);
-
-    // Check for the absence of 'Author Categories'
-    const authorCategories = await page
-      .locator("text='Author Categories'")
-      .count();
-    expect(authorCategories).toBe(0);
-  });
-  test("assert 'Marker Gene Sets' header is present and collapsed", async ({
-    page,
-  }) => {
-    await goToPage(page, pageURLCellGuide);
-
-    // Check for the presence of 'Marker Gene Sets' header
-    const markerGeneSetsHeader = await page.locator(
-      "h5:has-text('Marker Gene Sets')"
-    );
-    await expect(markerGeneSetsHeader).toBeVisible();
-
-    // Check if the header is collapsed by looking for a child with a chevron-right icon
-    const chevronRightIcon = markerGeneSetsHeader.locator(
-      "svg[data-icon='chevron-right']"
-    );
-    await expect(chevronRightIcon).toBeVisible();
-  });
-  test("expansion of 'Marker Gene Sets' reveals specific genesets", async ({
-    page,
-  }) => {
-    await goToPage(page, pageURLCellGuide);
-
-    // Locate and expand the 'Marker Gene Sets' header
-    const markerGeneSetsHeader = await page.locator(
-      "h5:has-text('Marker Gene Sets')"
-    );
-
-    tryUntil(
-      async () => {
-        await markerGeneSetsHeader.click(); // Assuming clicking will expand the section
-
-        // Assert the presence of specific genesets
-        const geneset1 = await page.locator(
-          `div[data-testid="geneset"]:has-text("enteric smooth muscle cell")`
-        );
-        const geneset2 = await page.locator(
-          `div[data-testid="geneset"]:has-text("smooth muscle fiber of ileum")`
-        );
-
-        await expect(geneset1).toBeVisible();
-        await expect(geneset2).toBeVisible();
-      },
-      { page }
-    );
-  });
-  test("expansion of 'enteric smooth muscle cell' reveals genes", async ({
-    page,
-  }) => {
-    await goToPage(page, pageURLCellGuide);
-
-    await expandMarkerGeneSetsHeader(page);
-
-    // Locate and expand 'enteric smooth muscle cell' geneset
-    await expandGeneset("enteric smooth muscle cell - marker genes", page);
-
-    // Assert the presence of the genes div and that it contains 55 child divs
-    const genesDiv = await page.locator(`div[data-testid="gene-set-genes"]`);
-    await expect(genesDiv).toBeVisible();
-    const childDivs = await genesDiv.locator(
-      'div[data-testid*=":gene-expand"]'
-    );
-    await expect(childDivs).toHaveCount(55);
-  });
-  test("deleting 'enteric smooth muscle cell' geneset and refreshing adds it back", async ({
-    page,
-  }) => {
-    await goToPage(page, pageURLCellGuide);
-
-    await expandMarkerGeneSetsHeader(page);
-
-    await deleteGeneset("enteric smooth muscle cell - marker genes", page);
-
-    // Refresh the page
-    await page.reload();
-
-    await expandMarkerGeneSetsHeader(page);
-
-    // Check if the geneset is added back
-    const genesetPresence = await page.locator(
-      `div[data-testid="geneset"]:has-text("enteric smooth muscle cell")`
-    );
-    await expect(genesetPresence).toBeVisible();
-  });
-});
 
 for (const testDataset of testDatasets) {
   const data = datasets[testDataset];
